+++ RELEASE_NOTES.md
@@ -4,8 +4,9 @@ This file contains one line summaries of commits that are worthy of mentioning i
 A new header is inserted each time a *tag* is created.
 
 ## main branch
- 
+- engine: Fix Adreno gpu crash introduced by gpu morph target change 
 - engine: Add optional memory configuration parameters to Engine initialization
+
 
 ## v1.25.2